Add: Added concurrent expression calculation
From my research, it looks like coroutines in Kotlin allow you to create "concurrent" behavior between different blocks of code in app. The idea was to keep UI on the main "thread" and use concurrency to move costly "expression calculation" to another thread. In that case there would be no delay in UI, hence -> Responsive UI. From the docs, it looks like viewModelScope is the most appropriate one as it keeps the thread running ONLY while viewmodel is "alive".