Rewrote backend...
Trying to get rid of Parsing at all possible stages. Instead of parsing of the string, I think, it would be better to implement enums. It's much easier to "add" new functions/operators/numbers to enums as needed and then add new logic then edit and trace all the string parsing in the code. So, code in that way should be more "scalable" and "maintainable".

Next:
Create a parsing interface that would be able to convert enums to string. that way I would not need to have strings all around the code and I will be able to "modify" linkage from enum to particular string representation of that enum in the one place.

This interface will be implemented by several classes (function/numbers/operators) or any other class that would need to have that conversion feature in the future.