Fix: representation of percentage in expression.
Due to the nature of my expression evaluator, the whole expression is getting evaluated at the same place (in the hope of getting some consistency and ease of maintainability). However, because of that, the way how some operations are portrayed cannot be changed until expression is evaluated. So, there was a bug where '%' was showing in expression, even though, by the design, it should not. This was fixed.

Bug: BigDecimal fills all numbers with trailing zeroes.... this behaviour leads to issues where I can't simply append digit to an existing number in my expression.
Bug: For some weird reason, float numbers cannot be "changed" in editing mode.