Rewrote logic for expression editing
Rewrote bunch of stuff in CalculatorFragment.kt to make editing feature more "maintainable"... Now it's dynamically assigns spannable to the new token, instead of reassignment of the all input on every new token encounter as it was before. Need to somehow make it more abstract in the future
Rewrote API for viewModel and Expression. Now it's less clunky and more straightforward (hopefully)